+++ src/main/java/fuzs/visualworkbench/element/VisualWorkbenchElement.java
@@ -9,14 +9,18 @@ import fuzs.visualworkbench.inventory.container.VisualWorkbenchContainer;
 import fuzs.visualworkbench.tileentity.WorkbenchTileEntity;
 import net.minecraft.block.Block;
 import net.minecraft.inventory.container.ContainerType;
+import net.minecraft.tags.BlockTags;
 import net.minecraft.tileentity.TileEntityType;
 import net.minecraft.util.ResourceLocation;
+import net.minecraftforge.common.Tags;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.ObjectHolder;
 
 public class VisualWorkbenchElement extends ClientExtensibleElement<VisualWorkbenchExtension> {
+
+    public static final Tags.IOptionalNamedTag<Block> NON_VISUAL_WORKBENCHES_TAG = BlockTags.createOptional(new ResourceLocation(VisualWorkbench.MODID, "non_visual_workbenches"));
 
     @ObjectHolder(VisualWorkbench.MODID + ":" + "crafting_table")
     public static final TileEntityType<WorkbenchTileEntity> WORKBENCH_TILE_ENTITY = null;
@@ -32,7 +36,7 @@ public class VisualWorkbenchElement extends ClientExtensibleElement<VisualWorkbe
     @Override
     public String[] getDescription() {
 
-        return new String[]{"Changing the in-game models to allow for better animations and subtle effects."};
+        return new String[]{"Items stay inside of crafting tables and are also rendered on top. It's really fancy!"};
     }
 
     @Override