+++ CHANGELOG.md
@@ -3,7 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
-## [v1.0] - 2021-07-01
+## [v1.0.1-1.16.5] - 2021-07-07
+#Added
+- Added `visualworkbench:non_visual_workbenches` block tag for blacklisting certain crafting tables 
+#Fixed
+- Fixed useless config file being generated
+- Rendering updates properly when using Crafting Tweaks mod
+
+## [v1.0-1.16.5] - 2021-07-03
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Puzzles Lib]: https://www.curseforge.com/minecraft/mc-mods/puzzles-lib